inicia fluxo realizado
