+++ utils/constants/general_constants.py
@@ -72,5 +72,53 @@ exceptions={
     "VILA OLIMPIA MASSAS": "Pagseguro Internet I.P. S.A.",
     "BATARD PADARIA ARTESANAL LTDA": "PAGAMENTO DE BOLETO - CELCOIN INSTITUICAO DE PAGAMEN",
     "TELEFONICA BRASIL S/A": ["PIX - ENVIADO - 16/07 14:57 CARLOS EDUARDO SANTOS DIB", "PIX - ENVIADO - 18/06 13:44 CARLOS EDUARDO SANTOS DIB", "PIX - ENVIADO - 21/05 11:26 CARLOS EDUARDO SANTOS DIB", "PIX - ENVIADO - 16/04 13:38 CARLOS EDUARDO SANTOS DIB", "PIX - ENVIADO - 26/03 15:30 CARLOS EDUARDO SANTOS DIB"]
-
 }  
+
+
+# Mapeamento Class_Cont_1 para Class_Cont_0
+mapeamento_class_cont = {
+    'Custo Mercadoria Vendida': 'Custo Mercadoria Vendida',
+    'Custos Artístico Geral': 'Custos Artístico Geral',
+    'Gorjeta': 'Mão de Obra',
+    'Utilidades': 'Utilidades',
+    'Mão de Obra - Salários': 'Mão de Obra',
+    'Impostos sobre Venda': 'Impostos',
+    'Custo de Ocupação': 'Ocupação',
+    'Mão de Obra - PJ': 'Mão de Obra',
+    'Mão de Obra - Extra': 'Mão de Obra',
+    'Mão de Obra - Benefícios': 'Mão de Obra',
+    'MAO DE OBRA FIXA/ TEMPORARIOS': 'Mão de Obra',
+    'Endividamento': 'Endividamento',
+    'CUSTO ARTISTICO': 'Custos Artístico Geral',
+    'Serviços de Terceiros': 'Serviços de Terceiros',
+    'Investimento - CAPEX': 'Investimento - Capex',
+    'Marketing': 'Marketing',
+    'INSUMOS': 'Custo Mercadoria Vendida',
+    'UTILIDADES': 'Utilidades',
+    'IMPOSTOS SOBRE VENDA': 'Impostos',
+    'Manutenção': 'Manutenção',
+    'Mão de Obra - Encargos e Provisões': 'Mão de Obra',
+    'ENDIVIDAMENTO': 'Endividamento',
+    'Dividendos e Remunerações Variáveis': 'Dividendos e Remunerações Variáveis',
+    'Locação de Equipamentos': 'Locação de Equipamentos',
+    'SERVICOS DE TERCEIROS': 'Serviços de Terceiros',
+    'CUSTOS COM MARKETING': 'Marketing',
+    'Informática e TI': 'Informática e TI',
+    'CUSTOS DE EVENTOS': 'Custos de Eventos',
+    'INVESTIMENTOS': 'Investimento - Capex',
+    'Despesas Financeiras': 'Despesas Financeiras',
+    'DESPESAS GERAIS': 'Utilidades',
+    'Custos de Eventos': 'Custos de Eventos',
+    'ADIANTAMENTO A FORNECEDORES': 'Custo Mercadoria Vendida',
+    'Deduções sobre Venda': 'Deduções sobre Venda',
+    'IMPOSTOS/ TRIBUTOS': 'Impostos',
+    'LOCACOES': 'Locação de Equipamentos',
+    'SISTEMAS/ T.I': 'Informática e TI',
+    'Despesas com Transporte / Hospedagem': 'Despesas com Transporte / Hospedagem',
+    'DESPESAS BANCARIAS': 'Despesas Financeiras',
+    'CUSTO DA MERCADORIA VENDIDA': 'Custo Mercadoria Vendida',
+    'CUSTO DE OCUPACAO': 'Ocupação',
+    'CUSTOS DE MANUTENCAO': 'Manutenção',
+    'DEDUCOES SOBRE VENDA': 'Deduções sobre Venda',
+    'Imposto de Renda': 'Impostos'
+}
