+++ utils/functions/general_functions.py
@@ -244,8 +244,11 @@ def mostrar_menu_permissoes_conciliacao(permissoes):
         st.sidebar.page_link("pages/Conciliação - Conciliações.py", label=":material/money_bag: Conciliação por casa")
         st.sidebar.page_link("pages/Conciliação - Farol_de_Conciliação.py", label=":material/finance: Farol de Conciliação")
         st.sidebar.page_link("pages/Conciliação - Ajustes.py", label=":material/instant_mix: Ajustes")
-        st.sidebar.write("")
-        st.sidebar.page_link("pages/Conciliação - Fluxo_de_Caixa.py", label=":material/currency_exchange: Fluxo de Caixa")
+        st.sidebar.markdown("## Fluxo de Caixa")
+        st.sidebar.page_link("pages/Conciliação - Fluxo_Realizado.py", label=":material/currency_exchange: Fluxo Realizado")
+        st.sidebar.page_link("pages/Conciliação - Fluxo_Futuro.py", label=":material/event_upcoming: Fluxo Futuro")
+
+    
 
 def config_sidebar():
